move: ApiResponseBody 및 구현체 패키지 이동
